Added `/team set-name` and `/team set-prefix`
