+++ plugin/src/main/java/io/github/wickeddroid/plugin/command/staff/CommandTeam.java
@@ -35,8 +35,6 @@ public class CommandTeam implements CommandClass {
   private Messages messages;
   @Inject
   private UhcGame uhcGame;
-  @Inject
-  private Teams teams;
 
   @Command(names = "enable")
   public void enable(
@@ -96,58 +94,5 @@ public class CommandTeam implements CommandClass {
           final Player leader
   ) {
     this.uhcTeamHandler.addPlayerToTeam(leader, target, true);
-  }
-
-  @Command(names = "set-name")
-  public void setName(final @Sender Player sender, @Named("name") String name) {
-    if(!teams.customization().allowCustomName()) {
-      messageHandler.send(sender, messages.team().changeTeamNameError());
-      return;
-    }
-
-    if(name.length() > teams.customization().maxNameLength()) {
-      messageHandler.send(sender, messages.team().maxLengthReached(), String.valueOf(teams.customization().maxNameLength()));
-      return;
-    }
-
-    var team = uhcTeamRegistry.getTeam(sender.getName());
-
-    if(team == null) {
-      messageHandler.send(sender, messages.team().playerDoesNotTeamExist());
-      return;
-    }
-
-    team.setName(name);
-
-    messageHandler.send(sender, messages.team().settingChanged(), name);
-  }
-
-  @Command(names = "set-prefix")
-  public void setPrefix(final @Sender Player sender, @Named("name") String prefix) {
-    if(!teams.customization().allowCustomPrefix()) {
-      messageHandler.send(sender, messages.team().changeTeamPrefixError());
-      return;
-    }
-
-    if(prefix.length() > teams.customization().maxPrefixLength()) {
-      messageHandler.send(sender, messages.team().maxLengthReached(), String.valueOf(teams.customization().maxPrefixLength()));
-      return;
-    }
-
-    var team = uhcTeamRegistry.getTeam(sender.getName());
-
-    if(team == null) {
-      messageHandler.send(sender, messages.team().playerDoesNotTeamExist());
-      return;
-    }
-
-    team.getTeam().prefix(MessageUtil.parseStringToComponent("["+prefix+"]")
-            .decoration(TextDecoration.OBFUSCATED, TextDecoration.State.FALSE)
-            .decoration(TextDecoration.ITALIC, TextDecoration.State.FALSE)
-            .decoration(TextDecoration.BOLD, TextDecoration.State.FALSE)
-            .decoration(TextDecoration.STRIKETHROUGH, TextDecoration.State.FALSE)
-            .decoration(TextDecoration.UNDERLINED, TextDecoration.State.FALSE));
-
-    messageHandler.send(sender, messages.team().settingChanged(), prefix);
   }
 }
